drivers: flash: flash_stm32_xspi: Allow flash driver init in ram load
Current driver protections prevents to initialize driver when run as an
external application when built into an application running from ext flash.
Aim is to avoid performing full driver initialization of the NOR flash
controller the application is read from.
But this is problem is actually only valid when application is running
in XIP mode (read in memory mapped mode at run time).
In ram load mode, since there is no direct activity from the application
to the NOR device, we can allow the controller initialization to happen,
at least partly. Mostly, what we need is to:
- skip the hal init
- abort memory mapping afterwards to let the jedec reading happen.

Note that this is required only for non memory mapped applications.

Signed-off-by: Erwan Gouriou <erwan.gouriou@st.com>